Moves defi to use axios tranformers
[ui tests]
