+++ frontend/app/src/services/defi/consts.ts
@@ -1,3 +1,5 @@
+import { balanceKeys } from '@/services/consts';
+
 export const DEFI_PROTOCOLS = ['aave', 'makerdao', 'compound'] as const;
 export const COMPOUND_EVENT_TYPES = [
   'mint',
@@ -7,3 +9,7 @@ export const COMPOUND_EVENT_TYPES = [
   'liquidation',
   'comp'
 ] as const;
+
+export const dsrKeys = [...balanceKeys, 'current_dsr'];
+export const vaultDetailsKeys = [...balanceKeys, 'total_interest_owed'];
+export const vaultKeys = [...balanceKeys, 'liquidation_price'];
